77 Wilson Gardens
"It's as well," said the old man; "it's a question whether I shall live to the Assizes, so it
matters little to me, but I should wish to spare Alice the shock. And now I will make the thing
clear to you; it has been a long time in the acting, but will not take me long to tell.

"You didn't know this dead man, McCarthy. He was a devil incarnate. I tell you that. God keep you
out of the clutches of such a man as he. His grip has been upon me these twenty years, and he has
blasted my life. I'll tell you first how I came to be in his power.