144 Glendale
"I came to find a friend."

"And I to find an enemy."

"An enemy?"

"Yes; one of my natural enemies, or, shall I say, my natural prey. Briefly, Watson, I am in the
midst of a very remarkable inquiry, and I have hoped to find a clue in the incoherent ramblings of
these sots, as I have done before now. Had I been recognised in that den my life would not have been
worth an hour's purchase; for I have used it before now for my own purposes, and the rascally Lascar
who runs it has sworn to have vengeance upon me. There is a trap-door at the back of that building,
near the corner of Paul's Wharf, which could tell some strange tales of what has passed through it
upon the moonless nights."

"What! You do not mean bodies?"