92 Down Close
"'I had £4 a month in my last place with Colonel Spence Munro.'

"'Oh, tut, tut! sweating--rank sweating!' he cried, throwing his fat hands out into the air like a
man who is in a boiling passion. 'How could anyone offer so pitiful a sum to a lady with such
attractions and accomplishments?'

"'My accomplishments, sir, may be less than you imagine,' said I. 'A little French, a little German,
music, and drawing--'

"'Tut, tut!' he cried. 'This is all quite beside the question. The point is, have you or have you
not the bearing and deportment of a lady? There it is in a nutshell. If you have not, you are not
fitted for the rearing of a child who may some day play a considerable part in the history of the
country. But if you have why, then, how could any gentleman ask you to condescend to accept anything
under the three figures? Your salary with me, madam, would commence at £100 a year.'