82 Glan Road
"I have devised seven separate explanations, each of which would cover the facts as far as we know
them. But which of these is correct can only be determined by the fresh information which we shall
no doubt find waiting for us. Well, there is the tower of the cathedral, and we shall soon learn all
that Miss Hunter has to tell."

The Black Swan is an inn of repute in the High Street, at no distance from the station, and there we
found the young lady waiting for us. She had engaged a sitting-room, and our lunch awaited us upon
the table.