149 Glan Road
Some three hours or so afterwards we were all in the train together, bound from Reading to the
little Berkshire village. There were Sherlock Holmes, the hydraulic engineer, Inspector Bradstreet,
of Scotland Yard, a plain-clothes man, and myself. Bradstreet had spread an ordnance map of the
county out upon the seat and was busy with his compasses drawing a circle with Eyford for its
centre.

"There you are," said he. "That circle is drawn at a radius of ten miles from the village. The place
we want must be somewhere near that line. You said ten miles, I think, sir."

"It was an hour's good drive."