45 Glan Road
"I am to remain neutral, to get near the window, to watch you, and at the signal to throw in this
object, then to raise the cry of fire, and to wait you at the corner of the street."

"Precisely."

"Then you may entirely rely on me."

"That is excellent. I think, perhaps, it is almost time that I prepare for the new role I have to
play."

He disappeared into his bedroom and returned in a few minutes in the character of an amiable and
simple-minded Nonconformist clergyman. His broad black hat, his baggy trousers, his white tie, his
sympathetic smile, and general look of peering and benevolent curiosity were such as Mr. John Hare
alone could have equalled. It was not merely that Holmes changed his costume. His expression, his
manner, his very soul seemed to vary with every fresh part that he assumed. The stage lost a fine
actor, even as science lost an acute reasoner, when he became a specialist in crime.