133 Wantage Close
"Well, there is at least a curious coincidence of dates. A ventilator is made, a cord is hung, and a
lady who sleeps in the bed dies. Does not that strike you?"

"I cannot as yet see any connection."

"Did you observe anything very peculiar about that bed?"

"No."

"It was clamped to the floor. Did you ever see a bed fastened like that before?"