56 Moor End
"He can't lie in the street. May we bring him in, marm?"

"Surely. Bring him into the sitting-room. There is a comfortable sofa. This way, please!"

Slowly and solemnly he was borne into Briony Lodge and laid out in the principal room, while I still
observed the proceedings from my post by the window. The lamps had been lit, but the blinds had not
been drawn, so that I could see Holmes as he lay upon the couch. I do not know whether he was seized
with compunction at that moment for the part he was playing, but I know that I never felt more
heartily ashamed of myself in my life than when I saw the beautiful creature against whom I was
conspiring, or the grace and kindliness with which she waited upon the injured man. And yet it would
be the blackest treachery to Holmes to draw back now from the part which he had intrusted to me. I
hardened my heart, and took the smoke-rocket from under my ulster. After all, I thought, we are not
injuring her. We are but preventing her from injuring another.