51 Beaconside
"My dear fellow, I know you well. I know the military neatness which characterises you. You shave
every morning, and in this season you shave by the sunlight; but since your shaving is less and less
complete as we get farther back on the left side, until it becomes positively slovenly as we get
round the angle of the jaw, it is surely very clear that that side is less illuminated than the
other. I could not imagine a man of your habits looking at himself in an equal light and being
satisfied with such a result. I only quote this as a trivial example of observation and inference.
Therein lies my métier, and it is just possible that it may be of some service in the investigation
which lies before us. There are one or two minor points which were brought out in the inquest, and
which are worth considering."

"What are they?"

"It appears that his arrest did not take place at once, but after the return to Hatherley Farm. On
the inspector of constabulary informing him that he was a prisoner, he remarked that he was not
surprised to hear it, and that it was no more than his deserts. This observation of his had the
natural effect of removing any traces of doubt which might have remained in the minds of the
coroner's jury."