95 Valentia Road
"The band! the speckled band!" whispered Holmes.

I took a step forward. In an instant his strange headgear began to move, and there reared itself
from among his hair the squat diamond-shaped head and puffed neck of a loathsome serpent.

"It is a swamp adder!" cried Holmes; "the deadliest snake in India. He has died within ten seconds
of being bitten. Violence does, in truth, recoil upon the violent, and the schemer falls into the
pit which he digs for another. Let us thrust this creature back into its den, and we can then remove
Miss Stoner to some place of shelter and let the county police know what has happened."

As he spoke he drew the dog-whip swiftly from the dead man's lap, and throwing the noose round the
reptile's neck he drew it from its horrid perch and, carrying it at arm's length, threw it into the
iron safe, which he closed upon it.

Such are the true facts of the death of Dr. Grimesby Roylott, of Stoke Moran. It is not necessary
that I should prolong a narrative which has already run to too great a length by telling how we
broke the sad news to the terrified girl, how we conveyed her by the morning train to the care of
her good aunt at Harrow, of how the slow process of official inquiry came to the conclusion that the
doctor met his fate while indiscreetly playing with a dangerous pet. The little which I had yet to
learn of the case was told me by Sherlock Holmes as we travelled back next day.

"I had," said he, "come to an entirely erroneous conclusion which shows, my dear Watson, how
dangerous it always is to reason from insufficient data. The presence of the gipsies, and the use of
the word 'band,' which was used by the poor girl, no doubt, to explain the appearance which she had
caught a hurried glimpse of by the light of her match, were sufficient to put me upon an entirely
wrong scent. I can only claim the merit that I instantly reconsidered my position when, however, it
became clear to me that whatever danger threatened an occupant of the room could not come either
from the window or the door. My attention was speedily drawn, as I have already remarked to you, to
this ventilator, and to the bell-rope which hung down to the bed. The discovery that this was a
dummy, and that the bed was clamped to the floor, instantly gave rise to the suspicion that the rope
was there as a bridge for something passing through the hole and coming to the bed. The idea of a
snake instantly occurred to me, and when I coupled it with my knowledge that the doctor was
furnished with a supply of creatures from India, I felt that I was probably on the right track. The
idea of using a form of poison which could not possibly be discovered by any chemical test was just
such a one as would occur to a clever and ruthless man who had had an Eastern training. The rapidity
with which such a poison would take effect would also, from his point of view, be an advantage. It
would be a sharp-eyed coroner, indeed, who could distinguish the two little dark punctures which
would show where the poison fangs had done their work. Then I thought of the whistle. Of course he
must recall the snake before the morning light revealed it to the victim. He had trained it,
probably by the use of the milk which we saw, to return to him when summoned. He would put it
through this ventilator at the hour that he thought best, with the certainty that it would crawl
down the rope and land on the bed. It might or might not bite the occupant, perhaps she might escape
every night for a week, but sooner or later she must fall a victim.