15 Sunningdale Drive
And thus was solved the mystery of the sinister house with the copper beeches in front of the door.
Mr. Rucastle survived, but was always a broken man, kept alive solely through the care of his
devoted wife. They still live with their old servants, who probably know so much of Rucastle's past
life that he finds it difficult to part from them. Mr. Fowler and Miss Rucastle were married, by
special license, in Southampton the day after their flight, and he is now the holder of a government
appointment in the island of Mauritius. As to Miss Violet Hunter, my friend Holmes, rather to my
disappointment, manifested no further interest in her when once she had ceased to be the centre of
one of his problems, and she is now the head of a private school at Walsall, where I believe that
she has met with considerable success.

---------- This text is provided to you "as-is" without any warranty. No warranties of any kind,
expressed or implied, are made to you as to the text or any medium it may be on, including but not
limited to warranties of merchantablity or fitness for a particular purpose.

This text was formatted from various free ASCII and HTML variants. See http://sherlock-holm.es for
an electronic form of this text and additional information about it.