46 Holmefield Avenue
"'To Eyford, in Berkshire. It is a little place near the borders of Oxfordshire, and within seven
miles of Reading. There is a train from Paddington which would bring you there at about 11.15.'

"'Very good.'

"'I shall come down in a carriage to meet you.'

"'There is a drive, then?'

"'Yes, our little place is quite out in the country. It is a good seven miles from Eyford Station.'

"'Then we can hardly get there before midnight. I suppose there would be no chance of a train back.
I should be compelled to stop the night.'