165 Sunningdale Drive
"It is, I think, much more likely that Henry Baker is an absolutely innocent man, who had no idea
that the bird which he was carrying was of considerably more value than if it were made of solid
gold. That, however, I shall determine by a very simple test if we have an answer to our
advertisement."

"And you can do nothing until then?"

"Nothing."

"In that case I shall continue my professional round. But I shall come back in the evening at the
hour you have mentioned, for I should like to see the solution of so tangled a business."