97 Valentia Road
"A certain selection and discretion must be used in producing a realistic effect," remarked Holmes.
"This is wanting in the police report, where more stress is laid, perhaps, upon the platitudes of
the magistrate than upon the details, which to an observer contain the vital essence of the whole
matter. Depend upon it, there is nothing so unnatural as the commonplace."

I smiled and shook my head. "I can quite understand your thinking so." I said. "Of course, in your
position of unofficial adviser and helper to everybody who is absolutely puzzled, throughout three
continents, you are brought in contact with all that is strange and bizarre. But here"--I picked up
the morning paper from the ground--"let us put it to a practical test. Here is the first heading
upon which I come. 'A husband's cruelty to his wife.' There is half a column of print, but I know
without reading it that it is all perfectly familiar to me. There is, of course, the other woman,
the drink, the push, the blow, the bruise, the sympathetic sister or landlady. The crudest of
writers could invent nothing more crude."

"Indeed, your example is an unfortunate one for your argument," said Holmes, taking the paper and
glancing his eye down it. "This is the Dundas separation case, and, as it happens, I was engaged in
clearing up some small points in connection with it. The husband was a teetotaler, there was no
other woman, and the conduct complained of was that he had drifted into the habit of winding up
every meal by taking out his false teeth and hurling them at his wife, which, you will allow, is not
an action likely to occur to the imagination of the average story-teller. Take a pinch of snuff,
Doctor, and acknowledge that I have scored over you in your example."

He held out his snuffbox of old gold, with a great amethyst in the centre of the lid. Its splendour
was in such contrast to his homely ways and simple life that I could not help commenting upon it.

"Ah," said he, "I forgot that I had not seen you for some weeks. It is a little souvenir from the
King of Bohemia in return for my assistance in the case of the Irene Adler papers."

"And the ring?" I asked, glancing at a remarkable brilliant which sparkled upon his finger.