189 Holmefield Avenue
"Not a bit, Doctor. Stay where you are. I am lost without my Boswell. And this promises to be
interesting. It would be a pity to miss it."

"But your client--"

"Never mind him. I may want your help, and so may he. Here he comes. Sit down in that armchair,
Doctor, and give us your best attention."

A slow and heavy step, which had been heard upon the stairs and in the passage, paused immediately
outside the door. Then there was a loud and authoritative tap.

"Come in!" said Holmes.

A man entered who could hardly have been less than six feet six inches in height, with the chest and
limbs of a Hercules. His dress was rich with a richness which would, in England, be looked upon as
akin to bad taste. Heavy bands of astrakhan were slashed across the sleeves and fronts of his
double-breasted coat, while the deep blue cloak which was thrown over his shoulders was lined with
flame-coloured silk and secured at the neck with a brooch which consisted of a single flaming beryl.
Boots which extended halfway up his calves, and which were trimmed at the tops with rich brown fur,
completed the impression of barbaric opulence which was suggested by his whole appearance. He
carried a broad-brimmed hat in his hand, while he wore across the upper part of his face, extending
down past the cheekbones, a black vizard mask, which he had apparently adjusted that very moment,
for his hand was still raised to it as he entered. From the lower part of the face he appeared to be
a man of strong character, with a thick, hanging lip, and a long, straight chin suggestive of
resolution pushed to the length of obstinacy.