84 Wilson Gardens
"Do you know the young lady?" I asked.

"Not I."

"It is half-past ten now."

"Yes, and I have no doubt that is her ring."

"It may turn out to be of more interest than you think. You remember that the affair of the blue
carbuncle, which appeared to be a mere whim at first, developed into a serious investigation. It may
be so in this case, also."