51 Harvey Street
"Yes," I answered, laughing. "It was a singular document. Philosophy, astronomy, and politics were
marked at zero, I remember. Botany variable, geology profound as regards the mud-stains from any
region within fifty miles of town, chemistry eccentric, anatomy unsystematic, sensational literature
and crime records unique, violin-player, boxer, swordsman, lawyer, and self-poisoner by cocaine and
tobacco. Those, I think, were the main points of my analysis."

Holmes grinned at the last item. "Well," he said, "I say now, as I said then, that a man should keep
his little brain-attic stocked with all the furniture that he is likely to use, and the rest he can
put away in the lumber-room of his library, where he can get it if he wants it. Now, for such a case
as the one which has been submitted to us to-night, we need certainly to muster all our resources.
Kindly hand me down the letter K of the 'American Encyclopaedia' which stands upon the shelf beside
you. Thank you. Now let us consider the situation and see what may be deduced from it. In the first
place, we may start with a strong presumption that Colonel Openshaw had some very strong reason for
leaving America. Men at his time of life do not change all their habits and exchange willingly the
charming climate of Florida for the lonely life of an English provincial town. His extreme love of
solitude in England suggests the idea that he was in fear of someone or something, so we may assume
as a working hypothesis that it was fear of someone or something which drove him from America. As to
what it was he feared, we can only deduce that by considering the formidable letters which were
received by himself and his successors. Did you remark the postmarks of those letters?"