17 Wellswood Gardens
"Pray continue," said Holmes. "Your narrative promises to be a most interesting one."

"You will find it rather disconnected, I fear, and there may prove to be little relation between the
different incidents of which I speak. On the very first day that I was at the Copper Beeches, Mr.
Rucastle took me to a small outhouse which stands near the kitchen door. As we approached it I heard
the sharp rattling of a chain, and the sound as of a large animal moving about.

"'Look in here!' said Mr. Rucastle, showing me a slit between two planks. 'Is he not a beauty?'

"I looked through and was conscious of two glowing eyes, and of a vague figure huddled up in the
darkness.