5 Ventnor Terrace
"There is at least a presumption that the vessel in which the man or men are is a sailing-ship. It
looks as if they always send their singular warning or token before them when starting upon their
mission. You see how quickly the deed followed the sign when it came from Dundee. If they had come
from Pondicherry in a steamer they would have arrived almost as soon as their letter. But, as a
matter of fact, seven weeks elapsed. I think that those seven weeks represented the difference
between the mail-boat which brought the letter and the sailing vessel which brought the writer."

"It is possible."

"More than that. It is probable. And now you see the deadly urgency of this new case, and why I
urged young Openshaw to caution. The blow has always fallen at the end of the time which it would
take the senders to travel the distance. But this one comes from London, and therefore we cannot
count upon delay."

"Good God!" I cried. "What can it mean, this relentless persecution?"

"The papers which Openshaw carried are obviously of vital importance to the person or persons in the
sailing-ship. I think that it is quite clear that there must be more than one of them. A single man
could not have carried out two deaths in such a way as to deceive a coroner's jury. There must have
been several in it, and they must have been men of resource and determination. Their papers they
mean to have, be the holder of them who it may. In this way you see K. K. K. ceases to be the
initials of an individual and becomes the badge of a society."