183 Holmefield Avenue
"But you are joking. What can you gather from this old battered felt?"

"Here is my lens. You know my methods. What can you gather yourself as to the individuality of the
man who has worn this article?"

I took the tattered object in my hands and turned it over rather ruefully. It was a very ordinary
black hat of the usual round shape, hard and much the worse for wear. The lining had been of red
silk, but was a good deal discoloured. There was no maker's name; but, as Holmes had remarked, the
initials "H. B." were scrawled upon one side. It was pierced in the brim for a hat-securer, but the
elastic was missing. For the rest, it was cracked, exceedingly dusty, and spotted in several places,
although there seemed to have been some attempt to hide the discoloured patches by smearing them
with ink.

"I can see nothing," said I, handing it back to my friend.