11 Glan Road
"You were not yourself at fault at all, then?"

"From the first, two facts were very obvious to me, the one that the lady had been quite willing to
undergo the wedding ceremony, the other that she had repented of it within a few minutes of
returning home. Obviously something had occurred during the morning, then, to cause her to change
her mind. What could that something be? She could not have spoken to anyone when she was out, for
she had been in the company of the bridegroom. Had she seen someone, then? If she had, it must be
someone from America because she had spent so short a time in this country that she could hardly
have allowed anyone to acquire so deep an influence over her that the mere sight of him would induce
her to change her plans so completely. You see we have already arrived, by a process of exclusion,
at the idea that she might have seen an American. Then who could this American be, and why should he
possess so much influence over her? It might be a lover; it might be a husband. Her young womanhood
had, I knew, been spent in rough scenes and under strange conditions. So far I had got before I ever
heard Lord St. Simon's narrative. When he told us of a man in a pew, of the change in the bride's
manner, of so transparent a device for obtaining a note as the dropping of a bouquet, of her resort
to her confidential maid, and of her very significant allusion to claim-jumping--which in miners'
parlance means taking possession of that which another person has a prior claim to--the whole
situation became absolutely clear. She had gone off with a man, and the man was either a lover or
was a previous husband--the chances being in favour of the latter."

"And how in the world did you find them?"

"It might have been difficult, but friend Lestrade held information in his hands the value of which
he did not himself know. The initials were, of course, of the highest importance, but more valuable
still was it to know that within a week he had settled his bill at one of the most select London
hotels."

"How did you deduce the select?"

"By the select prices. Eight shillings for a bed and eightpence for a glass of sherry pointed to one
of the most expensive hotels. There are not many in London which charge at that rate. In the second
one which I visited in Northumberland Avenue, I learned by an inspection of the book that Francis H.
Moulton, an American gentleman, had left only the day before, and on looking over the entries
against him, I came upon the very items which I had seen in the duplicate bill. His letters were to
be forwarded to 226 Gordon Square; so thither I travelled, and being fortunate enough to find the
loving couple at home, I ventured to give them some paternal advice and to point out to them that it
would be better in every way that they should make their position a little clearer both to the
general public and to Lord St. Simon in particular. I invited them to meet him here, and, as you
see, I made him keep the appointment."