17 Moor End
"Arthur does. Mary and I stay at home. We neither of us care for it."

"That is unusual in a young girl."

"She is of a quiet nature. Besides, she is not so very young. She is four-and-twenty."

"This matter, from what you say, seems to have been a shock to her also."