87 Sunningdale Drive
"I am sure that you are right, Mr. Holmes," cried our client. "A thousand things come back to me
which make me certain that you have hit it. Oh, let us lose not an instant in bringing help to this
poor creature."

"We must be circumspect, for we are dealing with a very cunning man. We can do nothing until seven
o'clock. At that hour we shall be with you, and it will not be long before we solve the mystery."

We were as good as our word, for it was just seven when we reached the Copper Beeches, having put up
our trap at a wayside public-house. The group of trees, with their dark leaves shining like
burnished metal in the light of the setting sun, were sufficient to mark the house even had Miss
Hunter not been standing smiling on the door-step.

"Have you managed it?" asked Holmes.

A loud thudding noise came from somewhere downstairs. "That is Mrs. Toller in the cellar," said she.
"Her husband lies snoring on the kitchen rug. Here are his keys, which are the duplicates of Mr.
Rucastle's."