2 Harvey Street
"Yes, I saw it by the side-lights when I was stepping into the carriage. It was a chestnut."

"Tired-looking or fresh?"

"Oh, fresh and glossy."

"Thank you. I am sorry to have interrupted you. Pray continue your most interesting statement."

"Away we went then, and we drove for at least an hour. Colonel Lysander Stark had said that it was
only seven miles, but I should think, from the rate that we seemed to go, and from the time that we
took, that it must have been nearer twelve. He sat at my side in silence all the time, and I was
aware, more than once when I glanced in his direction, that he was looking at me with great
intensity. The country roads seem to be not very good in that part of the world, for we lurched and
jolted terribly. I tried to look out of the windows to see something of where we were, but they were
made of frosted glass, and I could make out nothing save the occasional bright blur of a passing
light. Now and then I hazarded some remark to break the monotony of the journey, but the colonel
answered only in monosyllables, and the conversation soon flagged. At last, however, the bumping of
the road was exchanged for the crisp smoothness of a gravel-drive, and the carriage came to a stand.
Colonel Lysander Stark sprang out, and, as I followed after him, pulled me swiftly into a porch
which gaped in front of us. We stepped, as it were, right out of the carriage and into the hall, so
that I failed to catch the most fleeting glance of the front of the house. The instant that I had
crossed the threshold the door slammed heavily behind us, and I heard faintly the rattle of the
wheels as the carriage drove away.