52 Harvey Street
"But his voice was just a little too coaxing. He overdid it. I was keenly on my guard against him.

"'I was foolish enough to go into the empty wing,' I answered. 'But it is so lonely and eerie in
this dim light that I was frightened and ran out again. Oh, it is so dreadfully still in there!'