185 Glan Road
"None."

"One more question. How did your mother take the matter?"

"She was angry, and said that I was never to speak of the matter again."

"And your father? Did you tell him?"

"Yes; and he seemed to think, with me, that something had happened, and that I should hear of Hosmer
again. As he said, what interest could anyone have in bringing me to the doors of the church, and
then leaving me? Now, if he had borrowed my money, or if he had married me and got my money settled
on him, there might be some reason, but Hosmer was very independent about money and never would look
at a shilling of mine. And yet, what could have happened? And why could he not write? Oh, it drives
me half-mad to think of it, and I can't sleep a wink at night." She pulled a little handkerchief out
of her muff and began to sob heavily into it.