38 Bowler Street
"Oh, yes, with the deepest interest."

"That is well. The letter which I hold in my hand is from Lord St. Simon. I will read it to you, and
in return you must turn over these papers and let me have whatever bears upon the matter. This is
what he says: