142 Corbet Close
"I have ordered a carriage," said Lestrade as we sat over a cup of tea. "I knew your energetic
nature, and that you would not be happy until you had been on the scene of the crime."

"It was very nice and complimentary of you," Holmes answered. "It is entirely a question of
barometric pressure."

Lestrade looked startled. "I do not quite follow," he said.

"How is the glass? Twenty-nine, I see. No wind, and not a cloud in the sky. I have a caseful of
cigarettes here which need smoking, and the sofa is very much superior to the usual country hotel
abomination. I do not think that it is probable that I shall use the carriage to-night."

Lestrade laughed indulgently. "You have, no doubt, already formed your conclusions from the
newspapers," he said. "The case is as plain as a pikestaff, and the more one goes into it the
plainer it becomes. Still, of course, one can't refuse a lady, and such a very positive one, too.
She has heard of you, and would have your opinion, though I repeatedly told her that there was
nothing which you could do which I had not already done. Why, bless my soul! here is her carriage at
the door."