24 Sunningdale Drive
"'I was foolish enough to go into the empty wing,' I answered. 'But it is so lonely and eerie in
this dim light that I was frightened and ran out again. Oh, it is so dreadfully still in there!'

"'Only that?' said he, looking at me keenly.

"'Why, what did you think?' I asked.