72 Holmefield Avenue
"Precisely. You allude to my attempt to recover the Irene Adler papers, to the singular case of Miss
Mary Sutherland, and to the adventure of the man with the twisted lip. Well, I have no doubt that
this small matter will fall into the same innocent category. You know Peterson, the commissionaire?"

"Yes."

"It is to him that this trophy belongs."