Crime scene report #830268
"His son, you see, had grown up, and so had my girl, and as I was known to be in weak health, it
seemed a fine stroke to him that his lad should step into the whole property. But there I was firm.
I would not have his cursed stock mixed with mine; not that I had any dislike to the lad, but his
blood was in him, and that was enough. I stood firm. McCarthy threatened. I braved him to do his
worst. We were to meet at the pool midway between our houses to talk it over.

"When I went down there I found him talking with his son, so I smoked a cigar and waited behind a
tree until he should be alone. But as I listened to his talk all that was black and bitter in me
seemed to come uppermost. He was urging his son to marry my daughter with as little regard for what
she might think as if she were a slut from off the streets. It drove me mad to think that I and all
that I held most dear should be in the power of such a man as this. Could I not snap the bond? I was
already a dying and a desperate man. Though clear of mind and fairly strong of limb, I knew that my
own fate was sealed. But my memory and my girl! Both could be saved if I could but silence that foul
tongue. I did it, Mr. Holmes. I would do it again. Deeply as I have sinned, I have led a life of
martyrdom to atone for it. But that my girl should be entangled in the same meshes which held me was
more than I could suffer. I struck him down with no more compunction than if he had been some foul
and venomous beast. His cry brought back his son; but I had gained the cover of the wood, though I
was forced to go back to fetch the cloak which I had dropped in my flight. That is the true story,
gentlemen, of all that occurred."