Crime scene report #387351
"It is not yet nine. The streets will be crowded, so I trust that you may be in safety. And yet you
cannot guard yourself too closely."

"I am armed."

"That is well. To-morrow I shall set to work upon your case."

"I shall see you at Horsham, then?"

"No, your secret lies in London. It is there that I shall seek it."