Crime scene report #211147
"God help us!" said Holmes after a long silence. "Why does fate play such tricks with poor, helpless
worms? I never hear of such a case as this that I do not think of Baxter's words, and say, 'There,
but for the grace of God, goes Sherlock Holmes.'"

James McCarthy was acquitted at the Assizes on the strength of a number of objections which had been
drawn out by Holmes and submitted to the defending counsel. Old Turner lived for seven months after
our interview, but he is now dead; and there is every prospect that the son and daughter may come to
live happily together in ignorance of the black cloud which rests upon their past.