Crime scene report #136716
"Is such as we might expect. It ran, if I remember right, 'sent the pips to A, B, and C'--that is,
sent the society's warning to them. Then there are successive entries that A and B cleared, or left
the country, and finally that C was visited, with, I fear, a sinister result for C. Well, I think,
Doctor, that we may let some light into this dark place, and I believe that the only chance young
Openshaw has in the meantime is to do what I have told him. There is nothing more to be said or to
be done to-night, so hand me over my violin and let us try to forget for half an hour the miserable
weather and the still more miserable ways of our fellow-men."

It had cleared in the morning, and the sun was shining with a subdued brightness through the dim
veil which hangs over the great city. Sherlock Holmes was already at breakfast when I came down.

"You will excuse me for not waiting for you," said he; "I have, I foresee, a very busy day before me
in looking into this case of young Openshaw's."