Crime scene report #82038
"The letter arrived on March 10, 1883. His death was seven weeks later, upon the night of May 2nd."

"Thank you. Pray proceed."

"When my father took over the Horsham property, he, at my request, made a careful examination of the
attic, which had been always locked up. We found the brass box there, although its contents had been
destroyed. On the inside of the cover was a paper label, with the initials of K. K. K. repeated upon
it, and 'Letters, memoranda, receipts, and a register' written beneath. These, we presume, indicated
the nature of the papers which had been destroyed by Colonel Openshaw. For the rest, there was
nothing of much importance in the attic save a great many scattered papers and note-books bearing
upon my uncle's life in America. Some of them were of the war time and showed that he had done his
duty well and had borne the repute of a brave soldier. Others were of a date during the
reconstruction of the Southern states, and were mostly concerned with politics, for he had evidently
taken a strong part in opposing the carpet-bag politicians who had been sent down from the North.