Crime scene report #54409
"When Dr. Roylott was in India he married my mother, Mrs. Stoner, the young widow of Major-General
Stoner, of the Bengal Artillery. My sister Julia and I were twins, and we were only two years old at
the time of my mother's re-marriage. She had a considerable sum of money--not less than £1000 a year
--and this she bequeathed to Dr. Roylott entirely while we resided with him, with a provision that a
certain annual sum should be allowed to each of us in the event of our marriage. Shortly after our
return to England my mother died--she was killed eight years ago in a railway accident near Crewe.
Dr. Roylott then abandoned his attempts to establish himself in practice in London and took us to
live with him in the old ancestral house at Stoke Moran. The money which my mother had left was
enough for all our wants, and there seemed to be no obstacle to our happiness.

"But a terrible change came over our stepfather about this time. Instead of making friends and
exchanging visits with our neighbours, who had at first been overjoyed to see a Roylott of Stoke
Moran back in the old family seat, he shut himself up in his house and seldom came out save to
indulge in ferocious quarrels with whoever might cross his path. Violence of temper approaching to
mania has been hereditary in the men of the family, and in my stepfather's case it had, I believe,
been intensified by his long residence in the tropics. A series of disgraceful brawls took place,
two of which ended in the police-court, until at last he became the terror of the village, and the
folks would fly at his approach, for he is a man of immense strength, and absolutely uncontrollable
in his anger.

"Last week he hurled the local blacksmith over a parapet into a stream, and it was only by paying
over all the money which I could gather together that I was able to avert another public exposure.
He had no friends at all save the wandering gypsies, and he would give these vagabonds leave to
encamp upon the few acres of bramble-covered land which represent the family estate, and would
accept in return the hospitality of their tents, wandering away with them sometimes for weeks on
end. He has a passion also for Indian animals, which are sent over to him by a correspondent, and he
has at this moment a cheetah and a baboon, which wander freely over his grounds and are feared by
the villagers almost as much as their master.

"You can imagine from what I say that my poor sister Julia and I had no great pleasure in our lives.
No servant would stay with us, and for a long time we did all the work of the house. She was but
thirty at the time of her death, and yet her hair had already begun to whiten, even as mine has."