Crime scene report #532673
"Well, it is a possible supposition."

"You think so, too?"

"I did not say a probable one. But you do not yourself look upon this as likely?"

"I do not think Flora would hurt a fly."

"Still, jealousy is a strange transformer of characters. Pray what is your own theory as to what
took place?"

"Well, really, I came to seek a theory, not to propound one. I have given you all the facts. Since
you ask me, however, I may say that it has occurred to me as possible that the excitement of this
affair, the consciousness that she had made so immense a social stride, had the effect of causing
some little nervous disturbance in my wife."