Crime scene report #334783
"So much for the Lascar manager. Now for the sinister cripple who lives upon the second floor of the
opium den, and who was certainly the last human being whose eyes rested upon Neville St. Clair. His
name is Hugh Boone, and his hideous face is one which is familiar to every man who goes much to the
City. He is a professional beggar, though in order to avoid the police regulations he pretends to a
small trade in wax vestas. Some little distance down Threadneedle Street, upon the left-hand side,
there is, as you may have remarked, a small angle in the wall. Here it is that this creature takes
his daily seat, cross-legged with his tiny stock of matches on his lap, and as he is a piteous
spectacle a small rain of charity descends into the greasy leather cap which lies upon the pavement
beside him. I have watched the fellow more than once before ever I thought of making his
professional acquaintance, and I have been surprised at the harvest which he has reaped in a short
time. His appearance, you see, is so remarkable that no one can pass him without observing him. A
shock of orange hair, a pale face disfigured by a horrible scar, which, by its contraction, has
turned up the outer edge of his upper lip, a bulldog chin, and a pair of very penetrating dark eyes,
which present a singular contrast to the colour of his hair, all mark him out from amid the common
crowd of mendicants and so, too, does his wit, for he is ever ready with a reply to any piece of
chaff which may be thrown at him by the passers-by. This is the man whom we now learn to have been
the lodger at the opium den, and to have been the last man to see the gentleman of whom we are in
quest."

"But a cripple!" said I. "What could he have done single-handed against a man in the prime of life?"

"He is a cripple in the sense that he walks with a limp; but in other respects he appears to be a
powerful and well-nurtured man. Surely your medical experience would tell you, Watson, that weakness
in one limb is often compensated for by exceptional strength in the others."

"Pray continue your narrative."

"Mrs. St. Clair had fainted at the sight of the blood upon the window, and she was escorted home in
a cab by the police, as her presence could be of no help to them in their investigations. Inspector
Barton, who had charge of the case, made a very careful examination of the premises, but without
finding anything which threw any light upon the matter. One mistake had been made in not arresting
Boone instantly, as he was allowed some few minutes during which he might have communicated with his
friend the Lascar, but this fault was soon remedied, and he was seized and searched, without
anything being found which could incriminate him. There were, it is true, some blood-stains upon his
right shirt-sleeve, but he pointed to his ring-finger, which had been cut near the nail, and
explained that the bleeding came from there, adding that he had been to the window not long before,
and that the stains which had been observed there came doubtless from the same source. He denied
strenuously having ever seen Mr. Neville St. Clair and swore that the presence of the clothes in his
room was as much a mystery to him as to the police. As to Mrs. St. Clair's assertion that she had
actually seen her husband at the window, he declared that she must have been either mad or dreaming.
He was removed, loudly protesting, to the police-station, while the inspector remained upon the
premises in the hope that the ebbing tide might afford some fresh clue.