Crime scene report #25519
"He disappeared so suddenly."

"He might have leaped back. You did not see anyone else in the room?"

"No, but this horrible man confessed to having been there, and the Lascar was at the foot of the
stairs."

"Quite so. Your husband, as far as you could see, had his ordinary clothes on?"