Crime scene report #579262
"But it is impossible," said Miss Hunter; "the ladder was not there when the Rucastles went away."

"He has come back and done it. I tell you that he is a clever and dangerous man. I should not be
very much surprised if this were he whose step I hear now upon the stair. I think, Watson, that it
would be as well for you to have your pistol ready."