Crime scene report #600950
For a moment he had staggered and nearly fallen, but the brandy brought a tinge of colour into his
cheeks, and he sat staring with frightened eyes at his accuser.

"I have almost every link in my hands, and all the proofs which I could possibly need, so there is
little which you need tell me. Still, that little may as well be cleared up to make the case
complete. You had heard, Ryder, of this blue stone of the Countess of Morcar's?"

"It was Catherine Cusack who told me of it," said he in a crackling voice.