Crime scene report #60352
As he spoke the gleam of the side-lights of a carriage came round the curve of the avenue. It was a
smart little landau which rattled up to the door of Briony Lodge. As it pulled up, one of the
loafing men at the corner dashed forward to open the door in the hope of earning a copper, but was
elbowed away by another loafer, who had rushed up with the same intention. A fierce quarrel broke
out, which was increased by the two guardsmen, who took sides with one of the loungers, and by the
scissors-grinder, who was equally hot upon the other side. A blow was struck, and in an instant the
lady, who had stepped from her carriage, was the centre of a little knot of flushed and struggling
men, who struck savagely at each other with their fists and sticks. Holmes dashed into the crowd to
protect the lady; but just as he reached her he gave a cry and dropped to the ground, with the blood
running freely down his face. At his fall the guardsmen took to their heels in one direction and the
loungers in the other, while a number of better-dressed people, who had watched the scuffle without
taking part in it, crowded in to help the lady and to attend to the injured man. Irene Adler, as I
will still call her, had hurried up the steps; but she stood at the top with her superb figure
outlined against the lights of the hall, looking back into the street.

"Is the poor gentleman much hurt?" she asked.

"He is dead," cried several voices.

"No, no, there's life in him!" shouted another. "But he'll be gone before you can get him to
hospital."