Crime scene report #252544
There was a long silence, broken only by his heavy breathing and by the measured tapping of Sherlock
Holmes' finger-tips upon the edge of the table. Then my friend rose and threw open the door.

"Get out!" said he.

"What, sir! Oh, Heaven bless you!"

"No more words. Get out!"

And no more words were needed. There was a rush, a clatter upon the stairs, the bang of a door, and
the crisp rattle of running footfalls from the street.

"After all, Watson," said Holmes, reaching up his hand for his clay pipe, "I am not retained by the
police to supply their deficiencies. If Horner were in danger it would be another thing; but this
fellow will not appear against him, and the case must collapse. I suppose that I am commuting a
felony, but it is just possible that I am saving a soul. This fellow will not go wrong again; he is
too terribly frightened. Send him to jail now, and you make him a jail-bird for life. Besides, it is
the season of forgiveness. Chance has put in our way a most singular and whimsical problem, and its
solution is its own reward. If you will have the goodness to touch the bell, Doctor, we will begin
another investigation, in which, also a bird will be the chief feature."