Crime scene report #964913
"They have, however, allowed me a policeman, who may remain in the house with me."

"Has he come with you to-night?"

"No. His orders were to stay in the house."

Again Holmes raved in the air.

"Why did you come to me," he cried, "and, above all, why did you not come at once?"

"I did not know. It was only to-day that I spoke to Major Prendergast about my troubles and was
advised by him to come to you."