Crime scene report #575122
"It was naturally my intention that he should succeed me in my business, but he was not of a
business turn. He was wild, wayward, and, to speak the truth, I could not trust him in the handling
of large sums of money. When he was young he became a member of an aristocratic club, and there,
having charming manners, he was soon the intimate of a number of men with long purses and expensive
habits. He learned to play heavily at cards and to squander money on the turf, until he had again
and again to come to me and implore me to give him an advance upon his allowance, that he might
settle his debts of honour. He tried more than once to break away from the dangerous company which
he was keeping, but each time the influence of his friend, Sir George Burnwell, was enough to draw
him back again.

"And, indeed, I could not wonder that such a man as Sir George Burnwell should gain an influence
over him, for he has frequently brought him to my house, and I have found myself that I could hardly
resist the fascination of his manner. He is older than Arthur, a man of the world to his finger-
tips, one who had been everywhere, seen everything, a brilliant talker, and a man of great personal
beauty. Yet when I think of him in cold blood, far away from the glamour of his presence, I am
convinced from his cynical speech and the look which I have caught in his eyes that he is one who
should be deeply distrusted. So I think, and so, too, thinks my little Mary, who has a woman's quick
insight into character.